Major Refactoring
-Restructuring files.
-Removing CustomOperand, using only double instead.
-Including more operations.

+++ Formula/Source/Formula/FormulaBuilder.h
@@ -471,9 +471,9 @@ private:
                     return;
                 }
                 
-                float Number = StringTo<float>(NumberBuffer, NumberIndex);
+                double Number = StringTo<double>(NumberBuffer, NumberIndex);
                 
-                PushIntoData(OutParsingContext, OutData, OutDataIndex, EParticleType::Float32);
+                PushIntoData(OutParsingContext, OutData, OutDataIndex, EParticleType::Float64);
                 PushIntoData(OutParsingContext, OutData, OutDataIndex, Number);
 
                 OutParsingContext.Stage = EParsingStage::AfterOperand;
